requirement.txt
streamlit
spacy
pdfminer.six
